+++ src/finam/modules/debug.py
@@ -96,8 +96,7 @@ class DebugConsumer(TimeComponent):
         """dict[str, data] : The component's input data from the last time step"""
         return self._data
 
-    @property
-    def next_time(self):
+    def _next_time(self):
         return self.time + self._step
 
     def _initialize(self):